+++ Commands/CmdMegaboid.cs
@@ -101,7 +101,7 @@ namespace MCForge.Commands
             ushort b = p.level.GetTile(x, y, z);
             p.SendBlockchange(x, y, z, b);
             CatchPos cpos = (CatchPos)p.blockchangeObject;
-            unchecked { if (cpos.type != (byte)-1) type = cpos.type; else type = p.bindings[type]; }
+            unchecked { if (cpos.type != (byte)-1) type = cpos.type; else type = p.bindings[(int)type]; }
             List<Pos> buffer = new List<Pos>();
 
             ushort xx; ushort yy; ushort zz;
